back/requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
sqlparse==0.5.3
typing_extensions==4.12.2
requests==2.32.3
xrpl-py==4.0.0
djangorestframework-simplejwt
django-cors-headers==4.6.0
psycopg2-binary
dj-database-url
python-decouple==3.8
